refactor: reorganize test files
